Extraneous grass block generation in certain biomes when spawning meteors
I've often noticed that whenever meteors generate in other dimension or in certain biomes, they will generate with the grass as usual, this can look very odd in some cases and, albeit not very major, be a bit unpleasent.

This one simple if statement will, before attempting to generate grass, check if the biome's top block (in BiomeGenBase) is Blocks.grass. If it is, then generation continues as normal, if its not, then it will instead generate dirt, which is not as out of place in nearly all cases. It may be a good idea to add a config for the grass substitution in cases where dirt is out of place (it could instead, for example, be air or lava) however that's something that I felt should be discussed with you before it is implemented.

I've done testing with the wasteland mod biomes and stone shores/peaks.

This will not have any effect on performance and I feel will not be an addition that is disliked. I hope to see it added and understand if it isn't.

TL:DR

Grass will now only generate if the meteor generated in a biome where grass is the top block. 

This fixes some of the odd looking meteors when they spawn in certain biomes or dimensions.