Switched over to top block and filler block instead of a grass check.
Now, the grass and dirt generation is replaced with the generation of scattered top and filler blocks (For instance, grass and dirt). This means that the scattered debris around the meteorite will be based on the biome in which it crashed. 

This will work as long as the biome uses top/filler blocks. For modpack creators, a biome that doesn't use this would likely not be affected by biometweaker changes, meaning you could still set the top and filler blocks without ruining the biome.

This was tested in the badlands, wastelands (from the wasteland mod),  extreme hills, and deserts. Screenshots will be posted in the pull request