Created a config option to add Valid/Invalid Blocks to the list of possible spawn blocks that is found in MeteoritePlacer.java Added a note to the config to make the process as straight forward as possible. Includes many error messages to help config editors debug.
Added to the dimension whitelist config to allow overriding debris (helps it fit into the world even more)
Added a note to the config to make the process as straight forward as possible.

Modified registration.java to work with these changes, added a new method to get a block from the GameRegistry with a String input.

Modified MeteoriteBlockPutter.java and Fallout.java to correspond to these changes and generate blocks depending on the config.